Correct put_object semantics when passed a String
Previously, whenever a string was passed, it was always assumed
to be a file name. This has been corrected to be inline with
the pre-existing semantics of a passed data String being the
desired body.

Two new flavors of tests are introduced to object put operations,
one against File and one against String. Corresponding tests
are included on the create and set operations for fog File.

The some_temp_file_name test helper has been removed to prevent
confusion. In its place is some_temp_file, which returns a File
representing the contents at the path some_temp_file_name would
have returned.