Upgrade Compute addresses to >= 0.11
Also upgraded are the operation models and requests,
minimally to get the tests working.

This is a shift from my previous strategy of PRing requests
first then models for an entire module. That does not scale to
compute. For Compute, I'll PRing specific models as which will be
as implemented as possible, changing which requests are necessary.

This commit also proposes a change in the Compute test layout.
A flat directory, as tests are currently laid out in, makes it
difficult to match associated request and model tests.
This new layout introduces a directory per-model.

Note: there are several unimplemented portions of Address.
These are dependent on the server implementation and will land
when possible.